Move graphics::GraphicsContext to its own file
